implemented prediction methods + started viz
